Stop-gap measure for db_explain (#34)
The code that has been commented out will produce an output but it is
pretty badly mangled. I think this is because RJDBC::fetch coerces
resultset to be of type data.frame which may not be appropriate for the
resultset returned when SHOWPLAN is active. May need to write special
fetcher for this. But this will wait for future release of RSQLServer.